% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bah_data}
\alias{bah_data}
\title{BAH Data}
\format{
A data frame with 5 variables:
\describe{
  \item{MHA}{MHA code, ie "AK400"}
  \item{pg}{paygrade, ie "E01"}
  \item{dep}{dependency status, ie "With Dependents"}
  \item{year}
  \item{bah_rate}
}
}
\usage{
bah_data
}
\description{
A dataset containing BAH rates from 2008-2022
}
\keyword{datasets}
