% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_bah.R
\name{collect_bah}
\alias{collect_bah}
\title{Gather BAH data}
\usage{
collect_bah(
  mha,
  years = 2015:2022,
  dependency = "With Dependents",
  paygrades = "Limited"
)
}
\arguments{
\item{mha}{string of MHA code, ie AK400}

\item{years}{numeric series of years, default is 2015:2022}

\item{dependency}{dependency status, either "With Dependents" or "Without Dependents", default is "With Dependents"}

\item{paygrades}{either "Limited", which exludes enlisted and warrant officer or "All", default is "Limited"}
}
\value{
This function returns a dataframe of BAH rates
}
\description{
Gather BAH data
}
\author{
Bjorn Kallerud
}
