% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_mortality.R
\name{process_mortality}
\alias{process_mortality}
\title{Process Mortality Function}
\usage{
process_mortality(
  data,
  metadata_address,
  set_reference_year = NULL,
  var_maternal_mortality_cases = NULL
)
}
\arguments{
\item{data}{A data.table object containing the data to be processed.}

\item{metadata_address}{A string specifying the path to a JSON file containing metadata for processing the data.}

\item{set_reference_year}{(Optional) A numeric or string value specifying a reference year to be added to the data.}

\item{var_maternal_mortality_cases}{(Optional) A string specifying the name of the column in the data to be used for
identifying maternal mortality cases.}
}
\value{
Returns the processed data as a data.table object.
}
\description{
This function processes a given mortality dataset based on specified metadata json file,
and optional parameters for handling reference year, and maternal mortality cases.
}
\details{
The function consists of various helper functions that handle specific processing tasks:
\itemize{
\item \code{load_json_as_list}: Loads a JSON file and returns its content as a list.
\item \code{clean_string}: Cleans and standardizes character strings.
\item \code{check_actual_type}: Checks and returns the actual type of a categorical variable.
\item \code{process_date}: Processes and formats date values.
}

The main function reads the metadata from the specified JSON file, iterates through each variable in the metadata,
and processes the data based on the metadata specifications.

For each variable in the metadata, the function performs the following tasks:
\itemize{
\item Checks if the variable exists in the data, skips the variable if not found.
\item Checks and adjusts the data type of the variable based on the metadata and actual data.
\item Processes character, date, time, numeric, and categorical variables based on specified rules and formats.
\item Optionally creates a new variable 'maternal_mortality_cases' based on specified criteria.
\item Optionally adds a 'reference_year' variable with a specified value.
}
}
\examples{
\dontrun{
library(data.table)
library(jsonlite)

# Assume `mortality_data` is your data and `metadata.json` is your metadata file path address
processed_data <- process_mortality(mortality_data, "metadata.json", set_reference_year = 2020,
                                     var_maternal_mortality_cases = "icd_code_column")
}

}
\seealso{
\code{\link[jsonlite]{fromJSON}}, \code{\link[stringi]{stri_trans_general}},
\code{\link[base]{as.Date}}, \code{\link[dplyr]{na_if}}
}
