% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_to_pg.R
\name{rank_to_pg}
\alias{rank_to_pg}
\title{Rank to Paygrades}
\usage{
rank_to_pg(df, service)
}
\arguments{
\item{df}{a dataframe of ranks and possibly services}

\item{service}{either "Multi", in which case columns "service" and "rank" must be present in df, or one of "Army", "Air Force", "Navy", "Coast Guard", in which case column rank must be present in df}
}
\value{
This function returns a dataframe with an additional column for paygrade
}
\description{
Rank to Paygrades
}
\author{
Bjorn Kallerud
}
