% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Read File Function}
\usage{
read_file(file_path, sheet = NULL, csv_encoding = "UTF-8")
}
\arguments{
\item{file_path}{A string specifying the path to the file to be read.}

\item{sheet}{(Optional) A numeric or character value specifying the sheet to be read (only applicable for XLS and XLSX files).}

\item{csv_encoding}{(Optional) A string specifying the encoding of the CSV file. Default is "UTF-8".}
}
\value{
Returns the data read from the file as a data.table object with processed column names and character data.
}
\description{
This function reads files from a specified path and processes them into a data.table object.
Supported file extensions include CSV, TXT, DBF, SAV, XLS, and XLSX.
}
\details{
The function first identifies the file extension and then reads the file accordingly using the appropriate
R package and function. It handles missing values by replacing a set of predefined strings (e.g., "NA", "NULL") with NA.
Post-reading processing includes replacing certain characters, trimming whitespace, and standardizing column names.

The file reading is handled by different packages based on the file extension:
\itemize{
\item CSV and TXT files are read using \code{data.table::fread}.
\item DBF files are read using \code{foreign::read.dbf}.
\item SAV files are read using \code{foreign::read.spss}.
\item XLS and XLSX files are read using \code{readxl::read_excel}.
}
}
\examples{
\dontrun{
# Assume `file_path` is the path to your file
data <- read_file(file_path, sheet = 1)
}

}
\seealso{
\code{\link[data.table]{fread}}, \code{\link[foreign]{read.dbf}}, \code{\link[foreign]{read.spss}}, \code{\link[readxl]{read_excel}}
}
